# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alpha Deck",
    "summary": "Decision-support call-assembler API (advisory only; the CallCard is recomputed on read).",
    "version": "0.0.0"
  },
  "paths": {
...
    "/scoreboard": {
      "get": {
        "tags": [
          "scoreboard"
        ],
        "summary": "Get Scoreboard",
        "description": "The Scoreboard (SCORE): the call-of-record scored as-of ``asof`` \u2014 a READ-ONLY pass over\nthe immutable calls log, the operator-decision log, and realized asof-capped prices. The\nRECORD is the scoring source, never a recompute (replay re-derives history; this holds the\nplatform to what it actually said). Aggregate metrics judge only matured, non-censored\nepisodes and gate below ``min_n`` \u2014 an instrument, not a claim, until n accrues.",
        "operationId": "get_scoreboard_scoreboard_get",
        "parameters": [
          {
            "name": "asof",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "score the record as-of this date (caps both axes)",
              "title": "Asof"
            },
            "description": "score the record as-of this date (caps both axes)"
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "archived theses ride the record by default (archiving stops accrual, never erases the record); false is the explicit, reversible filter",
              "default": true,
              "title": "Include Archived"
            },
            "description": "archived theses ride the record by default (archiving stops accrual, never erases the record); false is the explicit, reversible filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/scoreboard/price-window": {
      "get": {
        "tags": [
          "scoreboard"
        ],
        "summary": "Get Price Window",
        "description": "One episode's realized daily OHLCV bars over ``[start, end]``, for the Scoreboard drawer's\nsparkline (Slice 3) \u2014 the SAME asof-capped read the scorer runs (``PgRealizedPrices``; ``bars_between``\nshares ``closes_between``'s cap/known_at), served on demand instead of embedded in the ledger payload\n(which stays lean). The line draws ``close``; open/high/low/volume ride the wire for a later candlestick.\n\nNo-lookahead (invariant #1) is enforced SERVER-SIDE and never trusted to the client: the reader caps\nthe valid-time axis at ``cap = asof`` (``d <= asof``), so a client passing a future ``end`` still gets\nonly bars ``<= asof`` \u2014 the window can never be widened past the as-of. ``known_at`` is left at the\nreader's default (now), matching the live scorer's own construction (``derive_thesis_record`` passes\n``known_at=None``), so the sparkline shows exactly the realized bars the ledger's numbers were computed\nfrom. Prices read under the THESIS'S tenant; a thesis the deployment tenant can't see is a 404.\n``source`` names the fact table the bars came from (invariant #6). ``start``/``end``/``security_id``\nride as bound params \u2014 the SQL range fragment stays a trusted literal (as ``closes_between`` documents).",
        "operationId": "get_price_window_scoreboard_price_window_get",
        "parameters": [
          {
            "name": "security_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "the episode's basket-member security to chart",
              "title": "Security Id"
            },
            "description": "the episode's basket-member security to chart"
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "window start \u2014 the episode's arm_date",
              "title": "Start"
            },
            "description": "window start \u2014 the episode's arm_date"
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "window end \u2014 the episode's exit_by; the series is still capped at asof server-side, so a future exit_by never widens it past asof",
              "title": "End"
            },
            "description": "window end \u2014 the episode's exit_by; the series is still capped at asof server-side, so a future exit_by never widens it past asof"
          },
          {
            "name": "asof",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "score as-of this date \u2014 the series is CAPPED here (no-lookahead)",
              "title": "Asof"
            },
            "description": "score as-of this date \u2014 the series is CAPPED here (no-lookahead)"
          },
          {
            "name": "thesis_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thesis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreboardPriceWindowOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PriceBar": {
        "properties": {
          "d": {
            "type": "string",
            "format": "date",
            "title": "D"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "d",
          "close"
        ],
        "title": "PriceBar",
        "description": "One realized daily OHLCV bar in an episode's price window (the sparkline draws ``close``; the\nfull bar rides the wire so a later candlestick is a pure-frontend swap \u2014 no second contract change).\n``close`` is non-null (null-close rows are skipped); ``open``/``high``/``low``/``volume`` are nullable\nper-column \u2014 a close-only free-EOD bar surfaces them as ``null``, never invented (invariant #6)."
      },
...
      "ScoreboardMetricOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "claim": {
            "type": "string",
            "title": "Claim"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "insufficient_n": {
            "type": "boolean",
            "title": "Insufficient N"
          },
          "summary": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Summary",
            "default": {}
          },
          "detail": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Detail",
            "default": []
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "claim",
          "n",
          "insufficient_n"
        ],
        "title": "ScoreboardMetricOut",
        "description": "One claim-tied metric (the replay set, computed over eligible live outcomes). ``claim``\nnames which system claim it tests \u2014 never a generic hit-rate; below ``n``/``insufficient_n``\nthe summary must not be read as a claim (the FE renders it quiet)."
      },
      "ScoreboardPriceWindowOut": {
        "properties": {
          "thesis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thesis Id"
          },
          "security_id": {
            "type": "string",
            "format": "uuid",
            "title": "Security Id"
          },
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          },
          "asof": {
            "type": "string",
            "format": "date",
            "title": "Asof"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "bars": {
            "items": {
              "$ref": "#/components/schemas/PriceBar"
            },
            "type": "array",
            "title": "Bars",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "thesis_id",
          "security_id",
          "start",
          "end",
          "asof",
          "source"
        ],
        "title": "ScoreboardPriceWindowOut",
        "description": "One episode's realized daily OHLCV series over ``[start, end]``, CAPPED at ``asof`` server-side \u2014\nthe drawer sparkline's on-demand read (Slice 3). It is the SAME asof-capped window the scorer runs\n(``PgRealizedPrices`` \u2014 ``bars_between`` shares ``closes_between``'s cap/known_at), exposed on request\nrather than embedded in the ledger payload. Invariant #1: no bar with ``d > asof`` is ever returned,\nwhatever ``end`` the client passes. ``source`` names the fact table the bars came from (invariant #6)."
      },
...
      "ScoreboardResponse": {
        "properties": {
          "asof": {
            "type": "string",
            "format": "date",
            "title": "Asof"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "summary": {
            "$ref": "#/components/schemas/ScoreboardSummaryOut"
          },
          "theses": {
            "items": {
              "$ref": "#/components/schemas/ScoreboardThesisOut"
            },
            "type": "array",
            "title": "Theses",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "asof",
          "generated_at",
          "summary"
        ],
        "title": "ScoreboardResponse",
        "description": "The Scoreboard: the call-of-record scored as-of ``asof`` \u2014 the record, never a recompute."
      },
      "ScoreboardSummaryOut": {
        "properties": {
          "n_theses": {
            "type": "integer",
            "title": "N Theses"
          },
          "n_with_record": {
            "type": "integer",
            "title": "N With Record"
          },
          "n_episodes": {
            "type": "integer",
            "title": "N Episodes"
          },
          "n_open": {
            "type": "integer",
            "title": "N Open"
          },
          "n_matured": {
            "type": "integer",
            "title": "N Matured"
          },
          "n_censored": {
            "type": "integer",
            "title": "N Censored"
          },
          "n_ingest_flagged": {
            "type": "integer",
            "title": "N Ingest Flagged",
            "default": 0
          },
          "n_eligible": {
            "type": "integer",
            "title": "N Eligible"
          },
          "n_takes": {
            "type": "integer",
            "title": "N Takes",
            "default": 0
          },
          "n_passes": {
            "type": "integer",
            "title": "N Passes",
            "default": 0
          },
          "n_overrides": {
            "type": "integer",
            "title": "N Overrides",
            "default": 0
          },
          "n_voided": {
            "type": "integer",
            "title": "N Voided",
            "default": 0
          },
          "record_began": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Began"
          },
          "banner": {
            "type": "string",
            "title": "Banner"
          },
          "min_n": {
            "type": "integer",
            "title": "Min N"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/ScoreboardMetricOut"
            },
            "type": "array",
            "title": "Metrics",
            "default": []
          },
          "next_maturity": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Maturity"
          },
          "n_maturing_30d": {
            "type": "integer",
            "title": "N Maturing 30D",
            "default": 0
          },
          "projected_min_n_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Min N Date"
          },
          "record_edge": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Edge"
          },
          "expected_asof": {
            "type": "string",
            "format": "date",
            "title": "Expected Asof"
          },
          "days_behind": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Behind"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "today": {
            "type": "string",
            "format": "date",
            "title": "Today"
          }
        },
        "type": "object",
        "required": [
          "n_theses",
          "n_with_record",
          "n_episodes",
          "n_open",
          "n_matured",
          "n_censored",
          "n_eligible",
          "banner",
          "min_n",
          "record_edge",
          "expected_asof",
          "days_behind",
          "today"
        ],
        "title": "ScoreboardSummaryOut",
        "description": "The aggregate strip: counts + the honesty banner + the gated metric set."
      },
      "ScoreboardThesisOut": {
        "properties": {
          "thesis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thesis Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker"
          },
          "basket_size": {
            "type": "integer",
            "title": "Basket Size",
            "default": 0
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "first_call_asof": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Call Asof"
          },
          "last_call_asof": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Call Asof"
          },
          "current_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current State"
          },
          "current_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Verdict"
          },
          "warming_since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warming Since"
          },
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/ScoreboardEpisodeOut"
            },
            "type": "array",
            "title": "Episodes",
            "default": []
          },
          "operator_spans": {
            "items": {
              "$ref": "#/components/schemas/OperatorSpanOut"
            },
            "type": "array",
            "title": "Operator Spans",
            "default": []
          },
          "decision_anomaly": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Anomaly"
          },
          "record_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Error"
          }
        },
        "type": "object",
        "required": [
          "thesis_id",
          "name"
        ],
        "title": "ScoreboardThesisOut",
        "description": "One thesis's slice of the Scoreboard: record coverage + scored episodes. Present even at\nzero episodes \u2014 the record span and an accruing warming window ARE the honest launch state.\n``record_error`` surfaces an unreadable historical card (fault isolation), never a 500."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}